requirements.txt
asgiref==3.2.10
Django==3.0.3
Pillow==7.1.2
pycryptodome==3.9.7
pytz==2020.1
sqlparse==0.3.1
astroid==2.4.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
django-messages==0.6.0
django-registration-redux==2.7
django-user-messages==0.7.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pylint==2.5.0
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2020.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1

